+++ libs/eco-store/auth/feature/login/src/lib/eco-store-auth-login/eco-store-auth-login.component.ts
@@ -6,6 +6,7 @@ import { MatIconModule } from '@angular/material/icon';
 import { RouterLink } from '@angular/router';
 import { TranslateModule } from '@ngx-translate/core';
 import { AUTH_FORM_FACADE } from '@plastik/auth/entities';
+import { pocketBaseUserProfileStore } from '@plastik/auth/pocketbase/data-access';
 import { PocketBaseImageUrlPipe } from '@plastik/eco-store/shared/utils';
 import { ecoStoreTenantStore } from '@plastik/eco-store/tenant';
 import { SharedFormFeatureModule } from '@plastik/shared/form';
@@ -32,5 +33,6 @@ import { SharedImgContainerComponent } from '@plastik/shared/img-container';
 })
 export class EcoStoreAuthLoginComponent {
   protected readonly facade = inject(AUTH_FORM_FACADE);
+  protected readonly profileStore = inject(pocketBaseUserProfileStore);
   protected readonly tenantStore = inject(ecoStoreTenantStore);
 }