+++ libs/eco-store/orders/data-access/src/eco-store-orders.store.ts
@@ -1,13 +1,15 @@
 import { inject } from '@angular/core';
-import { signalStore, withMethods } from '@ngrx/signals';
+import { signalStore, withMethods, withProps } from '@ngrx/signals';
 import { initialGetListState, withPocketBaseCrud } from '@plastik/signal-state/pocketbase';
 
 import { Router } from '@angular/router';
 import { DataCrud } from '@plastik/core/api-base';
 import { ecoStoreCartStore } from '@plastik/eco-store/cart/data-access';
 import { EcoStoreOrder } from '@plastik/eco-store/entities';
+import { activityStore } from '@plastik/shared/activity/data-access';
 import { ListResult } from 'pocketbase';
 import { EcoStoreOrdersApiService } from './eco-store-orders-api.service';
+
 export type OrdersPocketBaseCrudState = DataCrud<EcoStoreOrder, ListResult<EcoStoreOrder>>;
 
 export const ecoStoreOrdersStore = signalStore(
@@ -16,28 +18,33 @@ export const ecoStoreOrdersStore = signalStore(
     featureName: 'orders',
     dataServiceType: EcoStoreOrdersApiService,
     customInitialState: {
-      paginationSizeOptions: [20, 50, 75],
+      paginationSizeOptions: [10, 20, 40],
       pagination: {
         page: 1,
-        perPage: 20,
+        perPage: 10,
       },
       sortOptions: {
         ...initialGetListState().sortOptions,
       },
       apiRequestDebounceTime: 0,
+      isLoading: false,
     },
   }),
+  withProps(() => ({
+    _cartStore: inject(ecoStoreCartStore),
+    _router: inject(Router),
+    _activityStore: inject(activityStore),
+  })),
 
   withMethods(store => {
-    const cartStore = inject(ecoStoreCartStore);
-    const router = inject(Router);
-
     return {
       async createOrder() {
-        const data = cartStore.toOrder();
+        store._activityStore.setActivity(true, 'cart.finish.creatingOrder');
+        const data = store._cartStore.toOrder();
         const newOrder = await store.create(data);
-        cartStore.resetCartAfterCheckout();
-        await router.navigate(['/comanda', newOrder.id]);
+        store._cartStore.resetCartAfterCheckout();
+        await store._router.navigate(['/comanda', newOrder.id]);
+        store._activityStore.setActivity(false);
       },
     };
   })