+++ libs/eco-store/orders/feature/src/lib/order-confirmation/eco-store-order-confirmation.routes.ts
@@ -1,0 +1,9 @@
+import { Route } from '@angular/router';
+import { OrderConfirmationComponent } from './order-confirmation.component';
+
+export const ecoStoreOrderConfirmationRoutes: Route[] = [
+  {
+    path: ':id',
+    component: OrderConfirmationComponent,
+  },
+];