+++ libs/eco-store/products/feature/list/src/lib/eco-store-products-feature.resolver.ts
@@ -1,14 +1,13 @@
-import { inject, Injector } from '@angular/core';
+import { inject } from '@angular/core';
 import { toObservable } from '@angular/core/rxjs-interop';
 import { ActivatedRouteSnapshot, ResolveFn } from '@angular/router';
 import { ecoStoreProductCategoriesStore } from '@plastik/eco-store/product-categories/data-access';
 import { ecoStoreProductsStore } from '@plastik/eco-store/products/data-access';
-import { combineLatest, filter, map, take, tap } from 'rxjs';
+import { filter, map, take } from 'rxjs';
 
 export const ecoStoreProductsResolver: ResolveFn<boolean> = (route: ActivatedRouteSnapshot) => {
   const productStore = inject(ecoStoreProductsStore);
   const categoriesStore = inject(ecoStoreProductCategoriesStore);
-  const injector = inject(Injector);
   const queryParams = route.queryParams;
 
   const categorySlug = route.paramMap.get('category') ?? null;
@@ -21,23 +20,9 @@ export const ecoStoreProductsResolver: ResolveFn<boolean> = (route: ActivatedRou
     ...(categoryObj ? { category: categoryObj.category } : {}),
   });
 
-  return combineLatest([
-    toObservable(categoriesStore.initiallyLoaded, { injector }).pipe(filter(Boolean)),
-    toObservable(productStore.initiallyLoaded, { injector }).pipe(filter(Boolean)),
-  ]).pipe(
+  return toObservable(categoriesStore.isLoading).pipe(
+    filter(isLoading => !isLoading),
     take(1),
-    tap(() => {
-      // If we didn't have the category object before, try to set it now
-      if (!categoryObj) {
-        const categoryObjAfterLoad = categoriesStore.findCategoryBySlug(categorySlug);
-        if (categoryObjAfterLoad) {
-          productStore.setParams({
-            ...queryParams,
-            category: categoryObjAfterLoad.category,
-          });
-        }
-      }
-    }),
     map(() => true)
   );
 };