+++ libs/shared/signal-state/data-access-pocketbase/src/lib/pocketbase.features.ts
@@ -111,7 +111,6 @@ export function withPocketBaseListFeature<
                       }),
                       {
                         count: result.totalItems,
-                        initiallyLoaded: true,
                         isLoading: false,
                       }
                     );